# docs/swagger.yaml
definitions:
  operation.CreateTaskOP:
    properties:
      description:
        maxLength: 1000
        type: string
      due_date:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      status:
        type: integer
    required:
    - name
    type: object
...
  operation.SearchTaskOP:
    properties:
      info:
        type: string
    type: object
  operation.UpdateTaskOP:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      status:
        type: integer
    required:
    - name
    type: object
...
  serializer.ResponseTask:
    properties:
      data:
        $ref: '#/definitions/serializer.Task'
      error:
        type: string
      msg:
        type: string
      status:
        type: integer
    type: object
...
  serializer.Task:
    properties:
      created_at:
        type: integer
      description:
        example: sleep
        type: string
      due_date:
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: eat
        type: string
      start_time:
        type: integer
      status:
        example: 0
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/team/search:
    post:
      consumes:
      - application/json
      parameters:
      - description: search task info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/operation.SearchTaskOP'
      - description: team id
        in: query
        name: team_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":200,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
      summary: Search for team tasks
  /api/v1/team/task:
    post:
      consumes:
      - application/json
      parameters:
      - description: task info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/operation.CreateTaskOP'
      - description: team id is required when task is a team task
        in: query
        name: team_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":200,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
      summary: Create Team Task
  /api/v1/team/task/:task_id:
    delete:
      consumes:
      - application/json
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: string
      - description: team id
        in: query
        name: team_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":200,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
      summary: Delete team task
    put:
      consumes:
      - application/json
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: string
      - description: task info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/operation.UpdateTaskOP'
      - description: team id
        in: query
        name: team_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":200,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
      summary: Update team task
  /api/v1/team/tasks:
    get:
      consumes:
      - application/json
      parameters:
      - description: created_at | id | due_date | status
        in: query
        name: Index
        type: string
      - description: asc | desc
        in: query
        name: order
        type: string
      - description: record limit
        in: query
        name: limit
        type: integer
      - description: start from
        in: query
        name: start
        type: integer
      - description: team id
        in: query
        name: team_id
        required: true
        type: integer
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":200,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
      summary: Get Team task list
...
  /api/v1/user/search:
    post:
      consumes:
      - application/json
      parameters:
      - description: search task info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/operation.SearchTaskOP'
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":200,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
      summary: Search for personal tasks
  /api/v1/user/task:
    post:
      consumes:
      - application/json
      parameters:
      - description: task info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/operation.CreateTaskOP'
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":200,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
      summary: Create Personal Task
  /api/v1/user/task/:task_id:
    get:
      consumes:
      - application/json
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: string
      - description: team id
        in: query
        name: team_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":200,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
      summary: Show detailed information of team task
    put:
      consumes:
      - application/json
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: string
      - description: task info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/operation.UpdateTaskOP'
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":200,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
      summary: Update personal task
  /api/v1/user/tasks:
    get:
      consumes:
      - application/json
      parameters:
      - description: create_at | id | due_date | status
        in: query
        name: Index
        type: string
      - description: asc | desc
        in: query
        name: order
        type: string
      - description: recode limit
        in: query
        name: limit
        type: integer
      - description: start from
        in: query
        name: start
        type: integer
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":200,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
      summary: Get Personal task list
...
  /user/task/:task_id:
    delete:
      consumes:
      - application/json
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":200,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTask'
      summary: Delete personal task
swagger: "2.0"